requirements.txt
google-adk==1.1.1
google-generativeai==0.8.5
litellm==1.71.3
openpyxl==3.1.5
pandas==2.2.3
